Refacto (quadproduction/issues#361)

+++ start.py
@@ -692,7 +692,7 @@ def _initialize_environment(openpype_version: OpenPypeVersion) -> None:
     os.environ["PYTHONPATH"] = os.pathsep.join(split_paths)
 
 
-def _install_and_initialize_version(openpype_version: OpenPypeVersion):
+def _install_and_initialize_version(openpype_version: OpenPypeVersion, delete_zip=True):
     if openpype_version.path.is_file():
         _print(">>> Extracting zip file ...")
         try:
@@ -702,7 +702,7 @@ def _install_and_initialize_version(openpype_version: OpenPypeVersion):
             sys.exit(1)
         else:
             # cleanup zip after extraction, we don't touch prod dir
-            if openpype_version not in OpenPypeVersion.get_remote_versions():
+            if delete_zip and openpype_version not in OpenPypeVersion.get_remote_versions():
                 os.unlink(openpype_version.path)
 
         openpype_version.path = version_path
@@ -1073,24 +1073,29 @@ def boot():
         _boot_print_versions(OPENPYPE_ROOT)
         sys.exit(0)
 
-    # Check existing versions
-    user_dir_version_exists = local_version in OpenPypeVersion.get_local_versions()
-    prod_dir_version_exists = local_version in OpenPypeVersion.get_remote_versions()
+    # ------------------------------------------------------------------------
+    # Ensure patch version of OpenPype is on the user/local dir
+    # ------------------------------------------------------------------------
+    curr_version = local_version
+    if isinstance(local_version, str):
+        curr_version = OpenPypeVersion(version=local_version)
+    user_dir_version = bootstrap.find_openpype_local_version(curr_version)
+    prod_dir_version = bootstrap.find_openpype_remote_version(curr_version)
     dev_mode = "python" in os.path.basename(sys.executable).lower()
 
     op_version_to_extract = None
 
-    _print(">>> OP Version Exists in User Dir: {}".format(user_dir_version_exists))
-    _print(">>> OP Version Exists in Prod Dir: {}".format(prod_dir_version_exists))
+    _print(">>> OP Version Exists in User(local) Dir: {}".format(bool(user_dir_version)))
+    _print(">>> OP Version Exists in Prod(remote) Dir: {}".format(bool(prod_dir_version)))
     _print(">>> Dev Mode Enabled: {}".format(dev_mode))
 
-    if prod_dir_version_exists and not user_dir_version_exists:
+    if prod_dir_version and not user_dir_version:
         # Need to copy the prod version into the correct user/artist directory
         # Get OpenPypeVersion() Object
-        op_version_to_extract = bootstrap.find_openpype_version_in_prod_dir(local_version)
+        op_version_to_extract = prod_dir_version
 
-    if not prod_dir_version_exists and dev_mode:
-        # There isn't an official version, and we are in developer mode
+    if not prod_dir_version and dev_mode:
+        # This isn't a released version, we are in developer mode
         # Generate the version, then copy it on the user/artist directory
 
         # Generate Zip
@@ -1098,7 +1103,7 @@ def boot():
         op_version_to_extract = bootstrap.create_version_from_live_code(data_dir=debug_dir)
 
     if op_version_to_extract:
-        _install_and_initialize_version(op_version_to_extract)
+        _install_and_initialize_version(op_version_to_extract, delete_zip=False)
 
     # ------------------------------------------------------------------------
     # Find OpenPype versions
